Tweak `clean: false` when there's no .git dir
If the checkout directory doesn't contain .git, previously the action
always erased the entire directory: now, it only does it if the clean
input is true (which is the default).

This means that files set-up _before_ actions/checkout is used do not
get erased before cloning and checkout the repository.